Add extent building to db store
